# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-Commerce RestAPI",
    "description": "My E-Commerce RestAPI Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Register"],
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Login"],
        "description": "Login user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to logged-in",
            "schema": {
              "$ref": "#/definitions/UserLogged"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User is logged-in",
            "schema": {
              "$ref": "#/definitions/UserLogged"
            }
          }
        }
      }
    },
...
    "/api/products/": {
      "post": {
        "tags": ["Products"],
        "summary": "Create new product in system",
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Product that we want to create",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New product is created",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          }
        }
      }
    },
...
    "/api/products/{id}": {
      "put": {
        "summary": "Update product with given ID",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of product that we want to find",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product is update",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "properties": {
        "fullname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UserLogged": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "Products": {
      "properties": {
        "name": {
          "type": "string"
        },
        "quantityAvailabe": {
          "type": "number"
        },
        "costUnity": {
          "type": "number"
        },
        "user": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "Enter the token with the Bearer prefix, eg: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVC"
    }
  },
  "security": [{ "bearerAuth": [] }]
}